update `leap` solution to version 1.6.0

+++ rust/leap/src/lib.rs
@@ -1,4 +1,4 @@
-pub fn is_leap_year(year: i32) -> bool {
+pub fn is_leap_year(year: u64) -> bool {
     match year {
         y if (y % 400 == 0) => true,
         y if (y % 100 == 0) => false,
